Add tags validation to a taggable model
Problem: If you a taggable model contains invalid tags, and you check if
the model is valid, it will return true. I.e.:

    module Tag
     class Sport < Tagtical::Tag
       self.possible_values = %w{boxing basketball tennis hockey footabll soccer}
     end
    end

    class User < ActiveRecord::Base
     acts_as_taggable :sports
    end

    @user = User.new(:name => "Bobby")
    @user.sport_list = ["chess"]
    @user.valid? # => true
    @user.save # => will raise an exception

Solution: I added a validation rule to a taggable model, now it will
build tags from tag_lists and check if they are valid before validation
of the taggable model. And if they are not valid, it will add errors to
the taggable model.

Testing: Unit tests