Fix assigning of non-registered tags with the same values
If the Taggable model is trying to create a tag with the same value as
already created, but not registered tag class, it will be able to
successfully assign that tag to that model.

Instead, it should create new tag with that value, and don't use
existing not registered tag classes.